+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/internal/MessageListScrollHelper.kt
@@ -166,6 +166,8 @@ internal class MessageListScrollHelper(
                                 }
                             }
                         }
+
+                        if (it > SCROLL_BUTTON_VISIBILITY_THRESHOLD) scrollButtonView.isVisible = true
                     }
             },
             HIGHLIGHT_MESSAGE_DELAY